+++ R/model_spatial.R
@@ -125,7 +125,7 @@
 }
 
 .check_model_inputs<-function(y, meteo) {
-  if(!inherits(y, "sf")) stop("'y' has to be an object of class 'sf'.")
+  if(!inherits(y, "sf")) stop("'sf' has to be an object of class 'sf'.")
   if(!all(c("elevation","slope","aspect") %in% names(y))) stop("Columns 'elevation', 'slope' and 'aspect' must be defined.")
   if(!("forest" %in% names(y))) stop("Column 'forest' must be defined.")
   if(!("soil" %in% names(y))) stop("Column 'soil' must be defined.")
@@ -421,7 +421,8 @@
 #' 
 #' @seealso 
 #' \code{\link{spwb}}, \code{\link{growth}}, \code{\link{fordyn}}, \code{\link{spwb_spatial_day}}, 
-#' \code{\link{simulation_summary}} , \code{\link{plot_summary}}, \code{\link{update_landscape}}
+#' \code{\link{simulation_summary}} , \code{\link{plot_summary}}, 
+#' \code{\link{initialize_landscape}}, \code{\link{update_landscape}}
 #' 
 #' @examples
 #' \dontrun{
@@ -493,4 +494,132 @@ fordyn_spatial<-function(sf, SpParams, meteo = NULL, local_control = defaultCont
                 CO2ByYear = CO2ByYear, keep_results = keep_results, 
                 management_function = management_function, summary_function = summary_function, summary_arguments = summary_arguments, 
                 parallelize = parallelize, num_cores = num_cores, chunk_size = chunk_size, progress = progress)
+}
+
+#' Initialization of model inputs for spatially-distributed forest stands
+#' 
+#' Initializes state for local models \code{\link{spwb}} or \code{\link{growth}}. 
+#' 
+#' @param x An object of class \code{\link{sf}} with the following columns:
+#'   \itemize{
+#'     \item{\code{geometry}: Spatial geometry.}
+#'     \item{\code{forest}: Objects of class \code{\link{forest}}.}
+#'     \item{\code{soil}: Objects of class \code{\link{soil}} or data frames of physical properties.}
+#'     \item{\code{land_cover_type}: Land cover type of each grid cell (values should be 'wildland' or 'agriculture').}
+#'     \item{\code{crop_factor}: Crop evapo-transpiration factor. Only required for 'agriculture' land cover type.}
+#'   }
+#' @param SpParams A data frame with species parameters (see \code{\link{SpParamsMED}}).
+#' @param local_control A list of control parameters (see \code{\link{defaultControl}}).
+#' @param model A string to indicate the model, either \code{"spwb"} or \code{"growth"}.
+#' @param replace Boolean flag to replace existing initialized states
+#' @param progress Boolean flag to display progress information.
+#' 
+#' @returns Replaces or adds column 'state' whose elements are \code{\link{spwbInput}} or \code{\link{growthInput}} objects 
+#' and returns the modified object of class 'sf'.
+#' 
+#' @details
+#' Initialization is dealt automatically when calling simulation functions \code{\link{spwb_spatial}},  \code{\link{growth_spatial}},
+#' \code{\link{spwb_spatial_day}} or \code{\link{growth_spatial_day}}. However, function \code{initialize_landscape}  allows separating initialization from model simulations.
+#' 
+#' @author Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
+#' 
+#' @seealso 
+#' \code{\link{spwb_spatial}}, \code{\link{spwb_spatial_day}}, 
+#' \code{\link{update_landscape}}
+#' 
+#' @examples
+#' \dontrun{
+#' # Load example landscape data
+#' data("example_ifn")
+#'   
+#' # Load example meteo data frame from package meteoland
+#' data("examplemeteo")
+#'   
+#' # Load default medfate parameters
+#' data("SpParamsMED")
+#'   
+#' # Initialize state for 'spwb' simulations
+#' example_ifn <- initialize_landscape(example_ifn, SpParamsMED, 
+#'                                     defaultControl(), model = "spwb")
+#' 
+#' }
+#' 
+#' @name initialize_landscape
+initialize_landscape<- function(x, SpParams, local_control, model = "spwb", replace = FALSE, progress = TRUE) {
+  match.arg(model, c("spwb", "growth"))
+  if(!inherits(x, "sf")) stop("'x' has to be an object of class 'sf'.")
+  if(!("forest" %in% names(x))) stop("Column 'forest' must be defined.")
+  if(!("soil" %in% names(x))) stop("Column 'soil' must be defined.")
+  forestlist = x$forest
+  soillist  = x$soil
+  n <- length(forestlist)
+  if("state" %in% names(x)) {
+    xlist  = x$state
+  } else {
+    xlist = vector("list",n)
+  }
+  if("land_cover_type" %in% names(x)) {
+    landcover <- x$land_cover_type
+  } else {
+    landcover <- rep("wildland", n)
+  }
+  if("crop_factor" %in% names(x)) {
+    cropfactor <- x$crop_factor
+  } else {
+    cropfactor <- rep(NA, n)
+  }
+  n = length(forestlist)
+  if(model %in% c("spwb", "growth")) {
+    init<-rep(FALSE, n)
+    for(i in 1:n) {
+      if(landcover[i] == "wildland") {
+        f = forestlist[[i]]
+        s = soillist[[i]]
+        if(inherits(f, "forest") && inherits(s, c("soil", "data.frame"))) {
+          init[i] = TRUE
+          x_i = xlist[[i]]
+          if(!replace) {
+            if(inherits(x_i,"spwbInput") && model=="spwb") init[i] = FALSE
+            if(inherits(x_i,"growthInput") && model=="growth") init[i] = FALSE
+          }
+        }
+      } else if(landcover[i] == "agriculture") {
+        if(inherits(soillist[[i]], c("soil","data.frame"))) {
+          init[i] <- TRUE
+        }
+      }
+    }
+    w_init = which(init)
+    if(length(w_init)>0) {
+      if(progress) { 
+        cli::cli_progress_step(paste0("Creating '", length(w_init), "' input objects."))
+      }
+      for(w in 1:length(w_init)) {
+        i = w_init[w]
+        s = soillist[[i]]
+        if(inherits(s, "data.frame")) {
+          s <- soil(s)
+        }
+        if(landcover[i] == "wildland") {
+          f = forestlist[[i]]
+          if(inherits(f, "forest") && inherits(s, "soil")) {
+            if(model=="spwb") {
+              xlist[[i]] = medfate::forest2spwbInput(f, s, SpParams, local_control)
+            } else if(model=="growth") {
+              xlist[[i]] = medfate::forest2growthInput(f, s, SpParams, local_control)
+            }
+          }
+        } else if(landcover[i] == "agriculture") {
+          xlist[[i]] <- .aspwbInput(crop_factor = cropfactor[i], control = local_control, soil = s)
+        }
+      }
+      if(progress) {
+        cli::cli_progress_done()
+      }
+    } else {
+      if(progress) cli::cli_alert_info(paste0("All input objects are already available for '", model, "'."))
+    }
+  } 
+  x[["state"]] <- xlist
+  return(x)
 }